refactor: Moving endpoint resolution to HTTP invoker module
