+++ gravitee-gateway-core/src/main/java/io/gravitee/gateway/core/reactor/handler/impl/ApiHandlerConfiguration.java
@@ -15,7 +15,6 @@
  */
 package io.gravitee.gateway.core.reactor.handler.impl;
 
-import io.gravitee.gateway.core.endpoint.EndpointResolverFactory;
 import io.gravitee.gateway.core.http.spring.HttpClientConfiguration;
 import io.gravitee.gateway.core.policy.PathResolver;
 import io.gravitee.gateway.core.policy.PolicyConfigurationFactory;
@@ -23,6 +22,7 @@ import io.gravitee.gateway.core.policy.PolicyFactory;
 import io.gravitee.gateway.core.policy.PolicyResolver;
 import io.gravitee.gateway.core.policy.impl.*;
 import io.gravitee.gateway.core.reactor.handler.ContextReactorHandler;
+import io.gravitee.gateway.http.core.endpoint.EndpointResolverFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
